代码整理

+++ src/chapter06/kernel/dosfs.c
@@ -1332,7 +1332,7 @@ static int fat_open(struct fs *this, char *name, int mode, struct file **_fpp);
 static int fat_close  (struct file *_fp);
 static int fat_read   (struct file *_fp, uint8_t *buf, size_t size);
 static int fat_write  (struct file *_fp, uint8_t *buf, size_t size);
-static int fat_seek   (struct file *_fp, int offset, int whence);
+static int fat_seek   (struct file *_fp, off_t offset, int whence);
 
 struct fat_file {
   struct file file;
@@ -1468,7 +1468,7 @@ static int fat_write  (struct file *_fp, uint8_t *buf, size_t size)
 	return -1;
 }
 
-static int fat_seek   (struct file *_fp, int offset, int whence)
+static int fat_seek   (struct file *_fp, off_t offset, int whence)
 {
 	struct fat_file *fp = (struct fat_file *)_fp;
